refactor(server,types): extract type definition as template literal types break Github colorization
