+++ packages/temporal/src/types.ts
@@ -1,0 +1,13 @@
+export * from './workflows';
+
+export type ElectronicSignatureProcedureStatus =
+    | `PENDING.${
+          | 'WAITING_AGREEMENT'
+          | 'WAITING_EMAIL'
+          | 'GENERATING_CONFIRMATION_CODE'
+          | 'SENDING_CONFIRMATION_CODE'
+          | 'WAITING_CONFIRMATION_CODE'
+          | 'SIGNING_DOCUMENT'}`
+    | 'EXPIRED'
+    | 'VALIDATED'
+    | 'CANCELLED';
